feat: re-export CollectionTtl under cache namespace
