+++ src/lib.rs
@@ -23,7 +23,7 @@ pub use self::errors::*;
 pub use crate::credential_provider::CredentialProvider;
 pub use crate::response::simple_cache_client_sorted_set;
 pub use crate::simple_cache_client::{
-    CollectionTtl, Fields, IntoBytes, SimpleCacheClient, SimpleCacheClientBuilder,
+    Fields, IntoBytes, SimpleCacheClient, SimpleCacheClientBuilder,
 };
 
 pub use crate::cache_client::CacheClient;
