+++ src/simple_cache_client.rs
@@ -16,7 +16,7 @@ use std::ops::RangeBounds;
 use std::time::{Duration, UNIX_EPOCH};
 use tonic::{codegen::InterceptedService, transport::Channel, Request};
 
-use crate::response::simple_cache_client_sorted_set::{Elements, Order, Range, SortedSet};
+use crate::{cache::CollectionTtl, response::simple_cache_client_sorted_set::{Elements, Order, Range, SortedSet}};
 use crate::response::{
     DictionaryFetch, DictionaryGet, DictionaryPairs, Get, GetValue, ListCacheEntry, MomentoCache,
     MomentoCreateSigningKeyResponse, MomentoDeleteResponse, MomentoDictionaryDeleteResponse,
@@ -46,97 +46,6 @@ where
 {
     fn into_bytes(self) -> Vec<u8> {
         self.into()
-    }
-}
-
-/// Represents the desired behavior for managing the TTL on collection objects.
-///
-/// For cache operations that modify a collection (dictionaries, lists, or sets), there
-/// are a few things to consider. The first time the collection is created, we need to
-/// set a TTL on it. For subsequent operations that modify the collection you may choose
-/// to update the TTL in order to prolong the life of the cached collection object, or
-/// you may choose to leave the TTL unmodified in order to ensure that the collection
-/// expires at the original TTL.
-///
-/// The default behaviour is to refresh the TTL (to prolong the life of the collection)
-/// each time it is written using the client's default item TTL.
-#[derive(Copy, Clone, Debug)]
-pub struct CollectionTtl {
-    ttl: Option<Duration>,
-    refresh: bool,
-}
-
-impl CollectionTtl {
-    /// Create a collection TTL with the provided `ttl` and `refresh` settings.
-    pub const fn new(ttl: Option<Duration>, refresh: bool) -> Self {
-        Self { ttl, refresh }
-    }
-
-    /// Create a collection TTL that updates the TTL for the collection any time it is
-    /// modified.
-    ///
-    /// If `ttl` is `None` then the default item TTL for the client will be used.
-    pub fn refresh_on_update(ttl: impl Into<Option<Duration>>) -> Self {
-        Self::new(ttl.into(), true)
-    }
-
-    /// Create a collection TTL that will not refresh the TTL for the collection when
-    /// it is updated.
-    ///
-    /// Use this if you want to be sure that the collection expires at the originally
-    /// specified time, even if you make modifications to the value of the collection.
-    ///
-    /// The TTL will still be used when a new collection is created. If `ttl` is `None`
-    /// then the default item TTL for the client will be used.
-    pub fn initialize_only(ttl: impl Into<Option<Duration>>) -> Self {
-        Self::new(ttl.into(), false)
-    }
-
-    /// Create a collection TTL that updates the TTL for the collection only if an
-    /// explicit `ttl` is provided here.
-    pub fn refresh_if_provided(ttl: impl Into<Option<Duration>>) -> Self {
-        let ttl = ttl.into();
-        Self::new(ttl, ttl.is_some())
-    }
-
-    /// Return a new collection TTL which uses the same TTL but refreshes on updates.
-    pub fn with_refresh_on_update(self) -> Self {
-        Self::new(self.ttl(), true)
-    }
-
-    /// Return a new collection TTL which uses the same TTL but does not refresh on
-    /// updates.
-    pub fn with_no_refresh_on_update(self) -> Self {
-        Self::new(self.ttl(), false)
-    }
-
-    /// Return a new collecton TTL which has the same refresh behaviour but uses the
-    /// provided TTL.
-    pub fn with_ttl(self, ttl: impl Into<Option<Duration>>) -> Self {
-        Self::new(ttl.into(), self.refresh())
-    }
-
-    /// The [`Duration`] after which the cached collection should be expired from the
-    /// cache.
-    ///
-    /// If `None`, the default item TTL for the client will be used.
-    pub fn ttl(&self) -> Option<Duration> {
-        self.ttl
-    }
-
-    /// Whether the collection's TTL will be refreshed on every update.
-    ///
-    /// If true, this will extend the time at which the collection would expire when
-    /// an update operation happens. Otherwise, the collection's TTL will only be set
-    /// when it is initially created.
-    pub fn refresh(&self) -> bool {
-        self.refresh
-    }
-}
-
-impl Default for CollectionTtl {
-    fn default() -> Self {
-        Self::new(None, true)
     }
 }
 
@@ -679,7 +588,8 @@ impl SimpleCacheClient {
     /// use std::time::Duration;
     /// use std::iter::FromIterator;
     /// use std::collections::HashMap;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -742,7 +652,8 @@ impl SimpleCacheClient {
     /// use std::iter::FromIterator;
     /// use std::collections::HashMap;
     /// use std::convert::TryInto;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -818,7 +729,8 @@ impl SimpleCacheClient {
     /// use std::time::Duration;
     /// use std::iter::FromIterator;
     /// use std::collections::HashMap;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder, response::DictionaryFetch};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::{SimpleCacheClientBuilder, response::DictionaryFetch};
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -895,7 +807,8 @@ impl SimpleCacheClient {
     /// use std::time::Duration;
     /// use std::iter::FromIterator;
     /// use std::collections::HashMap;
-    /// use momento::{CollectionTtl, Fields, SimpleCacheClientBuilder, response::DictionaryFetch};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::{Fields, SimpleCacheClientBuilder, response::DictionaryFetch};
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -969,7 +882,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1032,7 +946,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1097,7 +1012,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1164,7 +1080,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1229,7 +1146,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1290,7 +1208,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1341,7 +1260,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1395,7 +1315,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1454,7 +1375,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1514,7 +1436,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1567,7 +1490,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1609,7 +1533,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1688,7 +1613,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1797,7 +1723,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -1854,7 +1781,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     /// use momento::response::MomentoSetDifferenceResponse;
     ///
     /// let ttl = CollectionTtl::default();
@@ -1972,7 +1900,8 @@ impl SimpleCacheClient {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::convert::TryInto;
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     /// use momento::response::simple_cache_client_sorted_set::{Elements, Order, SortedSetElement};
     ///
     /// let ttl = CollectionTtl::default();
@@ -2150,7 +2079,8 @@ impl SimpleCacheClient {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::convert::TryInto;
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     /// use momento::response::simple_cache_client_sorted_set::{Elements, Order, SortedSetElement};
     ///
     /// let ttl = CollectionTtl::default();
@@ -2498,7 +2428,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     ///
     /// let ttl = CollectionTtl::default();
     /// let mut momento = SimpleCacheClientBuilder::new(credential_provider, Duration::from_secs(30))?
@@ -2554,7 +2485,8 @@ impl SimpleCacheClient {
     /// # fn main() -> momento_test_util::DoctestResult {
     /// # momento_test_util::doctest(|cache_name, credential_provider| async move {
     /// use std::time::Duration;
-    /// use momento::{CollectionTtl, SimpleCacheClientBuilder};
+    /// use momento::cache::CollectionTtl;
+    /// use momento::SimpleCacheClientBuilder;
     /// use momento::response::simple_cache_client_sorted_set::SortedSetElement;
     ///
     /// let ttl = CollectionTtl::default();
